changed name of UserSettingsPage to UserProfilePage, and added some basic styling
